api_app/requirements.txt
bcrypt==4.0.1
beanie==1.19.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.2
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
motor==3.1.1
multidict==6.0.2
passlib==1.7.4
Pillow==9.3.0
pyasn1==0.4.8
pydantic==1.10.2
pymongo==4.4.0
pyparsing==3.0.9
python-dotenv   ==0.21.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
toml==0.10.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.18.0
webencodings==0.5.1
websockets==10.4
yarl==1.8.1
python-multipart
